api/pyproject.toml
[project]
name = "idea-fork-server"
version = "0.1.0"
description = "Idea Fork API - AI-powered product idea generation platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.0",
    # Redis and RQ for background job management (worker cronjob polling)
    "rq>=1.16.0",
    "redis>=5.0.0",
    # Shared idea generation pipeline
    "idea-generator",
    # Authentication
    "pyjwt>=2.8.0",
    "httpx>=0.27.0",
]

[tool.uv.sources]
idea-generator = { path = "../packages/idea-generator", editable = true }

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "httpx>=0.27.0",
]
migration = [
    "alembic>=1.17.2",
    "greenlet>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
